sdl/macos: disable `SDL_CAMERA` which we probably don't need and requires macOS frameworks

+++ game/system/hid/display_manager.cpp
@@ -346,8 +346,12 @@ void DisplayManager::update_video_modes() {
       display_name_str = display_name;
     }
 
-    DisplayMode new_mode = {display_name_str, curr_mode->format,       curr_mode->w,
-                            curr_mode->h,     curr_mode->refresh_rate, orient};
+    DisplayMode new_mode = {display_name_str,
+                            curr_mode->format,
+                            curr_mode->w,
+                            curr_mode->h,
+                            (int)curr_mode->refresh_rate,
+                            orient};
     m_current_display_modes[display_id] = new_mode;
     lg::info(
         "[DISPLAY]: Found monitor {}, currently set to {}x{}@{}hz. Format: {}, Orientation: {}",
